src/azure-function/requirements.txt
azure-functions==1.21.3
azure-servicebus==7.12.3
